+++ lib/elegant.dart
@@ -2,5 +2,6 @@ library elegant;
 
 export 'src/errors.dart';
 export 'src/extensions.dart';
+export 'src/iterable_extension.dart';
 export 'src/iterables.dart';
 export 'src/result.dart';